requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
crispy-bootstrap5==2023.10
defusedxml==0.7.1
Django==4.2.7
django-bootstrap-icons==0.8.6
django-bootstrap5==23.3
django-crispy-forms==2.1
idna==3.4
Pillow==10.1.0
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0